Cleanup recipe before saving changes.
I determined that the UI data must be dirty,
i.e. we can't trim spaces and newlines immediately.
Later, I could add smarter checks such as formatting onBlur.

For now, the data can be dirty in-display, but the database only contains clean data.

I determined some invalid states for data:
1. Most strings should be trimmed
2. Empty strings are usually invalid fields. A recipe title (which non-nullable) is an exception.
3. A subRecipe with no data is invalid UNLESS it is the only subRecipe
4. All recipes have at least one subRecipe, even if that subRecipe is blank.

These rules may further evolve as I think further.

One other rule: each step away from the user should make the data cleaner.
All dirty data is a result of dealing with uninformed users,
which is to be expected.
Though it may be important to keep the dirty data,
the database needs to strive to order the data early to make it useful to the system.