Allow deleting recipes when empty.
It may make sense later to add a delete button with a warning dialogue,
but if the recipe is completely empty,
it doesn't make sense to save it.

It's odd that I have the delete recipe logic within the saveRecipe portion,
but it fits nicely and maintains our debouncing logic.

I expect in the future to need a greater debouncing setup:
one which allows debouncing between functions and cancelling/aborting queries and queued items.