Fix the Saving text within the recipe editor.
The text would flash because there was no way to run code only when the debouncer had no queued work.
The debouncer can now output its current state to a function.

As part of this, I realized an issue with this setup.
After the queued work completes, the debouncer should wait an additional WaitMs
before running again.
Essentailly, the debouncer needs to be updated
so queued work is pulled off of the queue,
the state changes back to Locked,
and that state is reverted WaitMs after the queued job ends.
Perhaps I need setInterval instead of setTimeout?